requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
tensorflow-cpu==2.10.0
Pillow==9.5.0
numpy==1.23.5
scikit-learn==1.0.2
joblib==1.5.1
gunicorn==21.2.0